Extracting reusable functionality from gps_rescue.c to gps_rescue_common.c
