+++ src/test/Makefile
@@ -44,6 +44,7 @@ arming_prevention_unittest_SRC := \
 		$(USER_DIR)/fc/rc_modes.c \
 		$(USER_DIR)/fc/runtime_config.c \
 		$(USER_DIR)/flight/autopilot.c \
+		$(USER_DIR)/flight/gps_rescue_common.c \
 		$(USER_DIR)/flight/gps_rescue.c
